Refactor : Changed server package name to `server`
With some refactoring
